Reorganize file structure for API
