+++ api/books/addNotes.ts
@@ -1,5 +1,5 @@
-import { authHeaders, baseUrl } from "./base";
-import { ListItem } from "./types";
+import { authHeaders, baseUrl } from "../base";
+import { ListItem } from "../types";
 
 export const addNotes = (notes: string, list: ListItem) => {
   fetch(`${baseUrl()}/api/v1/list_item/${list.id}`, {
