+++ api/books/removeFromReadingList.ts
@@ -1,4 +1,4 @@
-import { authHeaders, baseUrl } from "./base";
+import { authHeaders, baseUrl } from "../base";
 
 export const removeFromReadingList = (
   listID: number,
