+++ api/books/setRating.ts
@@ -1,4 +1,4 @@
-import { authHeaders, baseUrl } from "./base";
+import { authHeaders, baseUrl } from "../base";
 
 export const setRating = (id: number, rating: number) => {
   fetch(`${baseUrl()}/api/v1/list_item/${id}`, {
