+++ pages/discover.tsx
@@ -1,5 +1,5 @@
 import { useContext, useEffect } from "react";
-import { fetchBooks } from "../api/fetchBooks";
+import { fetchBooks } from "../api/books/fetchbooks";
 import { BookList } from "../components/BookList";
 import { Search } from "../components/Search";
 import { BooksContext } from "../contexts/BooksContext";
